+++ examples/getting-started/getting-started-project/src/components/WeatherCards.tsx
@@ -1,3 +1,4 @@
+import dayjs from 'dayjs';
 import weatherStyles from '../styles/weather.module.css';
 import { getWeatherIcon } from '../utils.ts';
 
@@ -20,6 +21,22 @@ export function CurrentWeatherCard({ weatherData }: any) {
       </div>
       <h2 className={weatherStyles.city}>{location.name}</h2>
       <span className={weatherStyles.country}>{location.country}</span>
+    </div>
+  );
+}
+
+export function UpcomingWeatherCard({ weatherData }: any) {
+  return (
+    <div className={weatherStyles.card}>
+      <span>{dayjs(weatherData.date).format('dddd')}</span>
+      <img
+        src={getWeatherIcon(weatherData.day.condition.text.trim())}
+        alt={weatherData.day.condition.text}
+      />
+      <h3>
+        {weatherData.day.avgtemp_f}&deg;
+        <span className={weatherStyles.unit}>F</span>
+      </h3>
     </div>
   );
 }
