+++ examples/getting-started/getting-started-project/src/components/modules/Footer/index.tsx
@@ -1,8 +1,19 @@
-export function Component() {
-  return <div>Footer</div>;
+import { ModuleFields, TextField } from '@hubspot/cms-components/fields';
+import footerStyles from '../../../styles/footer.module.css';
+
+export function Component({ fieldValues }: any) {
+  return (
+    <footer className={footerStyles.footer}>
+      <p>{fieldValues.footerText}</p>
+    </footer>
+  );
 }
 
-export const fields = [];
+export const fields = (
+  <ModuleFields>
+    <TextField label="Footer Text" name="footerText" default="Be Well." />
+  </ModuleFields>
+);
 
 export const meta = {
   label: 'Footer Module',
