+++ examples/getting-started/getting-started-project/src/styles/footer.module.css
@@ -1,0 +1,5 @@
+.footer {
+  text-align: center;
+  font-weight: bold;
+  font-size: 2em;
+}
