+++ examples/getting-started/getting-started-project/src/styles/header.module.css
@@ -1,8 +1,3 @@
-:root {
-  --primary-color: #ff7a59;
-  --accent-color: #2d3e50;
-}
-
 .wrapper {
   width: 100%;
   background-color: transparent;
