+++ examples/getting-started/getting-started-project/src/styles/weather.module.css
@@ -2,7 +2,6 @@
   --primary-color: #ff7a59;
   --primary-color-dark: #e66e50;
   --primary-color-med: #ffbcac;
-  --primary-color-light: #fff1ee;
   --secondary-color: #dfe3eb;
   --accent-color: #2d3e50;
   --light-accent-color: #eaf0f6;
