+++ examples/getting-started/getting-started-project/src/tsconfig.json
@@ -1,21 +1,18 @@
 {
   "compilerOptions": {
-    "target": "ES2017",
+    "target": "esnext",
     "module": "Node16",
-    "moduleResolution": "node16",
-    "strict": true,
+    "moduleResolution": "Node16",
     "esModuleInterop": true,
-    "skipLibCheck": true,
+    "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
     "allowImportingTsExtensions": true,
     "noEmit": true,
+    "isolatedModules": true,
     "jsx": "react-jsx",
-    "plugins": [
-      {
-        "name": "typescript-plugin-css-modules"
-      }
-    ]
+    "skipLibCheck": true,
+    "types": ["vite/client", "vitest/globals"]
   },
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Recommended"
+  "include": ["components/**/*"],
+  "exclude": ["node_modules"]
 }
