+++ examples/getting-started/getting-started-theme/weather.hubl.html
@@ -1,7 +1,7 @@
 <!--
   templateType: page
   isAvailableForNewContent: true
-  label: Hello World - Cars
+  label: Getting Started - Weather Forecast
 -->
 {% extends "./layouts/base.hubl.html" %}
